requirements.txt
# API
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.13.1

# LangChain
langchain==1.2.13
langchain-core==1.2.22
langchain-community==0.4.1
langchain-openai==1.1.12
langchain-groq==1.1.2
langchain-ollama==1.0.1
langchain-text-splitters==1.1.1
langgraph==1.1.3

# LLM clients
openai==2.30.0
groq==0.37.1

# Embeddings + reranking
sentence-transformers==5.3.0
torch==2.11.0
transformers==5.4.0

# Vector DB
weaviate-client==4.20.4

# PDF ingestion
PyMuPDF==1.27.2.2

# Evaluation
ragas==0.4.3

# Caching
redis==7.4.0

# Utilities
numpy==2.4.3
python-dotenv==1.2.2